+++ Runtime/AnimationJobs/IAnimatableProperty.cs
@@ -1,4 +1,4 @@
-﻿namespace UnityEngine.Animations.Rigging
+namespace UnityEngine.Animations.Rigging
 {
     using Experimental.Animations;
 
@@ -149,6 +149,82 @@
             x.SetFloat(stream, value.x);
             y.SetFloat(stream, value.y);
             z.SetFloat(stream, value.z);
+        }
+    }
+
+    public struct Vector3IntProperty : IAnimatableProperty<Vector3Int>
+    {
+        public PropertyStreamHandle x;
+        public PropertyStreamHandle y;
+        public PropertyStreamHandle z;
+
+        public static Vector3IntProperty Bind(Animator animator, Component component, string name)
+        {
+            var type = component.GetType();
+            return new Vector3IntProperty
+            {
+                x = animator.BindStreamProperty(component.transform, type, name + ".x"),
+                y = animator.BindStreamProperty(component.transform, type, name + ".y"),
+                z = animator.BindStreamProperty(component.transform, type, name + ".z")
+            };
+        }
+
+        public static Vector3IntProperty BindCustom(Animator animator, string name)
+        {
+            return new Vector3IntProperty
+            {
+                x = animator.BindCustomStreamProperty(name + ".x", CustomStreamPropertyType.Int),
+                y = animator.BindCustomStreamProperty(name + ".y", CustomStreamPropertyType.Int),
+                z = animator.BindCustomStreamProperty(name + ".z", CustomStreamPropertyType.Int)
+            };
+        }
+
+        public Vector3Int Get(AnimationStream stream) =>
+            new Vector3Int(x.GetInt(stream), y.GetInt(stream), z.GetInt(stream));
+
+        public void Set(AnimationStream stream, Vector3Int value)
+        {
+            x.SetInt(stream, value.x);
+            y.SetInt(stream, value.y);
+            z.SetInt(stream, value.z);
+        }
+    }
+
+    public struct Vector3BoolProperty : IAnimatableProperty<Vector3Bool>
+    {
+        public PropertyStreamHandle x;
+        public PropertyStreamHandle y;
+        public PropertyStreamHandle z;
+
+        public static Vector3BoolProperty Bind(Animator animator, Component component, string name)
+        {
+            var type = component.GetType();
+            return new Vector3BoolProperty
+            {
+                x = animator.BindStreamProperty(component.transform, type, name + ".x"),
+                y = animator.BindStreamProperty(component.transform, type, name + ".y"),
+                z = animator.BindStreamProperty(component.transform, type, name + ".z")
+            };
+        }
+
+        public static Vector3BoolProperty BindCustom(Animator animator, string name)
+        {
+            return new Vector3BoolProperty
+            {
+                x = animator.BindCustomStreamProperty(name + ".x", CustomStreamPropertyType.Bool),
+                y = animator.BindCustomStreamProperty(name + ".y", CustomStreamPropertyType.Bool),
+                z = animator.BindCustomStreamProperty(name + ".z", CustomStreamPropertyType.Bool)
+            };
+        }
+
+        public Vector3Bool Get(AnimationStream stream) =>
+            new Vector3Bool(x.GetBool(stream), y.GetBool(stream), z.GetBool(stream));
+
+        public void Set(AnimationStream stream, Vector3Bool value)
+        {
+            x.SetBool(stream, value.x);
+            y.SetBool(stream, value.y);
+            z.SetBool(stream, value.z);
         }
     }