+++ Runtime/AnimationRig/Constraints/MultiPositionConstraint.cs
@@ -53,6 +53,7 @@ namespace UnityEngine.Animations.Rigging
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Multi-Position Constraint")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/MultiPositionConstraint.html")]
     public class MultiPositionConstraint : RigConstraint<
         MultiPositionConstraintJob,
         MultiPositionConstraintData,