+++ Runtime/AnimationRig/Constraints/MultiReferentialConstraint.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 
 namespace UnityEngine.Animations.Rigging
 {
@@ -58,6 +58,7 @@ namespace UnityEngine.Animations.Rigging
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Multi-Referential Constraint")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/MultiReferentialConstraint.html")]
     public class MultiReferentialConstraint : RigConstraint<
         MultiReferentialConstraintJob,
         MultiReferentialConstraintData,