+++ Runtime/AnimationRig/Constraints/OverrideTransform.cs
@@ -1,4 +1,4 @@
-﻿namespace UnityEngine.Animations.Rigging
+namespace UnityEngine.Animations.Rigging
 {
     [System.Serializable]
     public struct OverrideTransformData : IAnimationJobData, IOverrideTransformData
@@ -50,6 +50,7 @@
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Override Transform")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/OverrideTransform.html")]
     public class OverrideTransform : RigConstraint<
         OverrideTransformJob,
         OverrideTransformData,