+++ Runtime/AnimationRig/Constraints/TwistCorrection.cs
@@ -58,6 +58,7 @@ namespace UnityEngine.Animations.Rigging
     }
 
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Twist Correction")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/constraints/TwistCorrection.html")]
     public class TwistCorrection : RigConstraint<
         TwistCorrectionJob,
         TwistCorrectionData,