+++ Runtime/AnimationRig/RigTransform.cs
@@ -1,6 +1,7 @@
-﻿namespace UnityEngine.Animations.Rigging
+namespace UnityEngine.Animations.Rigging
 {
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Setup/Rig Transform")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@latest?preview=1&subfolder=/manual/index.html")]
     public class RigTransform : MonoBehaviour
     {
     }