+++ Runtime/Utils/QuaternionExt.cs
@@ -2,6 +2,10 @@ namespace UnityEngine.Animations.Rigging
 {
     public static class QuaternionExt
     {
+        const float k_FloatMin = 1e-10f;
+
+        public static readonly Quaternion zero = new Quaternion(0f, 0f, 0f, 0f);
+
         public static Quaternion FromToRotation(Vector3 from, Vector3 to)
         {
             float theta = Vector3.Dot(from.normalized, to.normalized);
@@ -18,6 +22,29 @@ namespace UnityEngine.Animations.Rigging
             }
 
             return Quaternion.AngleAxis(Mathf.Acos(theta) * Mathf.Rad2Deg, Vector3.Cross(from, to).normalized);
+        }
+
+        public static Quaternion Add(Quaternion rhs, Quaternion lhs)
+        {
+            float sign = Mathf.Sign(Quaternion.Dot(rhs, lhs));
+            return new Quaternion(rhs.x + sign * lhs.x, rhs.y + sign * lhs.y, rhs.z + sign * lhs.z, rhs.w + sign * lhs.w);
+        }
+
+        public static Quaternion Scale(Quaternion q, float scale)
+        {
+            return new Quaternion(q.x * scale, q.y * scale, q.z * scale, q.w * scale);
+        }
+
+        public static Quaternion NormalizeSafe(Quaternion q)
+        {
+            float dot = Quaternion.Dot(q, q);
+            if (dot > k_FloatMin)
+            {
+                float rsqrt = 1.0f / Mathf.Sqrt(dot);
+                return new Quaternion(q.x * rsqrt, q.y * rsqrt, q.z * rsqrt, q.w * rsqrt);
+            }
+
+            return Quaternion.identity;
         }
     }
 }